refactor(EmailFieldset): separate registration-specific logic into it's own wrapper.
