+++ src/components/Fieldsets/EmailFieldset.js
@@ -1,30 +1,15 @@
-/* global $IS_DEVELOPMENT, $IS_STAGING */
-
 import React, { forwardRef } from 'react'
 
-import InputFieldset, { useValidationCallback } from './InputFieldset'
+import InputFieldset from './InputFieldset'
 
 
-const isProduction = !$IS_DEVELOPMENT && !$IS_STAGING
+
 
 
 const EmailFieldset = forwardRef((props, ref) => {
   const {
-    onValidate: parentValidate,
     ...inputProps
   } = props
-
-  const handleValidate = useValidationCallback(
-    (messages, value) => {
-      if (value === 'surly_badger@fuelrats.com') {
-        messages.errors.push('While we appreciate your admiration for our founder, this email is most surely not your own.')
-      } else if (isProduction && value.includes('@fuelrats')) {
-        messages.errors.push('Fuel rats email accounts cannot be used to register a brand new account, goofball.')
-      }
-    },
-    [],
-    parentValidate,
-  )
 
   return (
     <InputFieldset
@@ -32,8 +17,7 @@ const EmailFieldset = forwardRef((props, ref) => {
       displayName="E-Mail"
       placeholder="surly_badger@fuelrats.com"
       {...inputProps}
-      type="email"
-      onValidate={handleValidate} />
+      type="email" />
   )
 })
 EmailFieldset.displayName = 'EmailFieldset'
